fastapi_app/requirements.txt
fastapi
uvicorn
pydantic[email]
sqlalchemy
pymysql
python-dotenv
cryptography
requests
passlib
bcrypt==3.2.2
python-jose[cryptography]
python-multipart